+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Previously changelog for versions `1.0`-`1.4` has been located in README.md. How
 ## SkyDocu `1.5` (__WIP__)
 - Added _Requested containers_ metric to superadministration Container statistics widget
 - Code cleanup
+- API endpoints
+- External systems
+    - Right management
+    - Logging
 
 ## SkyDocu `1.4` (03/22/2025)
 - In-container multiple database support